Renaming api mapper and parser
