+++ src/routes/api/notes/+server.ts
@@ -5,7 +5,7 @@ import { taskEither as TE } from 'fp-ts/lib';
 
 import { updateBoard, getBoardByUserId } from '$lib/server/db/boardDb';
 import { createNote } from '$lib/server/db/notesDb';
-import { parseRequest } from '$lib/server/parseRequest';
+import { parseRequest } from '$lib/server/requestParser';
 import { mapToApiError } from '$lib/server/apiResultMapper';
 import { CreateNoteInputSchema } from '$lib/types';
